refactor: Moves AddIDsToReplicationSpecs to TPF and reuse FormatMongoDBMajorVersion
